WIP: Feature/strip creation (#3)
* Add Name to strip

* Added strips.go

* Added config/strips.go

* Added strips-example.yaml

* Added 'name' to NewStrip

* Read stripDef and build strips from stripDef

* Change pin value from float32 to float64

* Use clamp for pwmPin.Set()

* Remove redundant Off() call

* Remove commented out code

* Add pubsub to glide dependencies

* Move strip creation to config

* Remove blah.go

* Remove f.Sync(). Cannot sync named pipe.

* Updated .gitignore to exclude vendor

+++ controller/rgb.go
@@ -2,7 +2,7 @@ package controller
 
 // RGB is a struct representing the Red, Blue and Green values
 type RGB struct {
-	Red   float32
-	Green float32
-	Blue  float32
+	Red   float64
+	Green float64
+	Blue  float64
 }
